+++ database/migrations/2024_11_04_131514_create_national_immunization_programs_table.php
@@ -11,33 +11,35 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('national_immunization_programs', function (Blueprint $table) {
-            $table->bigIncrements('immunizationId'); 
-            $table->unsignedBigInteger('personalId'); // Foreign key column
-            $table->unsignedBigInteger('id');
-            $table->string('birthplace', 100);
-            $table->string('bloodtype', 5);
-            $table->string('mothername', 100);
-            $table->string('dswdNhts', 100);
-            $table->string('facilityHouseholdno', 100);
-            $table->string('houseHoldno', 100);
-            $table->string('fourpsmember', 100);
-            $table->string('PCBMember', 100);
-            $table->string('philhealthStatus', 50);
-            $table->string('philhealthNo', 15);
-            $table->string('ifMember', 100);
-            $table->string('familyMember', 100);
-            $table->string('ttStatus', 10);
-            $table->date('dateAssesed');
-            $table->date('date');
-            $table->string('place', 100);
-            $table->string('guardian', 100);
-            $table->timestamps();
+        if (!Schema::hasTable('national_immunization_programs')) {
+            Schema::create('national_immunization_programs', function (Blueprint $table) {
+                $table->bigIncrements('immunizationId'); 
+                $table->unsignedBigInteger('personalId'); // Foreign key column
+                $table->unsignedBigInteger('id');
+                $table->string('birthplace', 100);
+                $table->string('bloodtype', 5);
+                $table->string('mothername', 100);
+                $table->string('dswdNhts', 100);
+                $table->string('facilityHouseholdno', 100);
+                $table->string('houseHoldno', 100);
+                $table->string('fourpsmember', 100);
+                $table->string('PCBMember', 100);
+                $table->string('philhealthStatus', 50);
+                $table->string('philhealthNo', 15);
+                $table->string('ifMember', 100);
+                $table->string('familyMember', 100);
+                $table->string('ttStatus', 10);
+                $table->date('dateAssesed');
+                $table->date('date');
+                $table->string('place', 100);
+                $table->string('guardian', 100);
+                $table->timestamps();
 
-            // Setting up the foreign key constraint
-            $table->foreign('personalId')->references('personalId')->on('personal_information')->onDelete('cascade');
-            $table->foreign('id')->references('id')->on('users')->onDelete('cascade');
-        });
+                // Setting up the foreign key constraint
+                $table->foreign('personalId')->references('personalId')->on('personal_information')->onDelete('cascade');
+                $table->foreign('id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
